fix(driver/modern_bpf): avoid type confusion setting snaplen

+++ driver/modern_bpf/programs/tail_called/events/syscall_dispatched_events/send.bpf.c
@@ -71,10 +71,10 @@ int BPF_PROG(send_x,
 	unsigned long args[3];
 	extract__network_args(args, 3, regs);
 
-	uint16_t bytes_to_read = ret > 0 ? ret : args[2];
+	int64_t bytes_to_read = ret > 0 ? ret : args[2];
 	uint16_t snaplen = maps__get_snaplen();
 	apply_dynamic_snaplen(regs, &snaplen, false);
-	if(snaplen > bytes_to_read)
+	if((int64_t)snaplen > bytes_to_read)
 	{
 		snaplen = bytes_to_read;
 	}
